Move UserAttributeController to User directory
- Created `app\Http\Controllers\User` directory.
- Moved `UserAttributeController.php` into the new `User` directory.
- Updated the namespace in `UserAttributeController.php` to `App\Http\Controllers\User`.
- Modified `routes/api.php` to reference the relocated `UserAttributeController`.
